+++ jinja2_simple_tags.py
@@ -1,61 +1,98 @@
+import warnings
+from typing import Any, ClassVar, Dict, List, Tuple
+
 from jinja2 import nodes
 from jinja2.ext import Extension
+from jinja2.lexer import describe_token
+from jinja2.parser import Parser
+from jinja2.runtime import Context
 
-__all__ = ['StandaloneTag', 'ContainerTag']
-__version__ = '0.4.1'
+__all__ = ["StandaloneTag", "ContainerTag"]
+__version__ = "0.4.1"
 
 
 class BaseTemplateTag(Extension):
     def __init__(self, environment):
         super().__init__(environment)
+        self.context = None
         self.template = None
         self.lineno = None
         self.tag_name = None
 
-    def parse(self, parser):
+    def parse(self, parser: Parser) -> nodes.Node:
         lineno = parser.stream.current.lineno
         tag_name = parser.stream.current.value
         additional_params = [
-            nodes.Keyword('_context', nodes.ContextReference()),
-            nodes.Keyword('_template', nodes.Const(parser.name)),
-            nodes.Keyword('_lineno', nodes.Const(lineno)),
-            nodes.Keyword('_tag_name', nodes.Const(tag_name)),
+            nodes.Keyword("_context", nodes.ContextReference()),
+            nodes.Keyword("_template", nodes.Const(parser.name)),
+            nodes.Keyword("_lineno", nodes.Const(lineno)),
+            nodes.Keyword("_tag_name", nodes.Const(tag_name)),
         ]
-        self.init_parser(parser)
-        args, kwargs, target = self.parse_args(parser)
-        kwargs.extend(additional_params)
-        call_node = self.call_method('render_wrapper', args, kwargs)
-        return self.output(parser, call_node, target, tag_name=tag_name, lineno=lineno)
 
-    def init_parser(self, parser):
+        self.init_parser(parser)
+        args, kwargs, options = self.parse_args(parser)
+        kwargs.extend(additional_params)
+        options.setdefault("tag_name", tag_name)
+
+        if hasattr(self, "output") and callable(self.output):
+            warnings.warn(
+                "The \"output\" method of the \"BaseTemplateTag\" class is deprecated "
+                "and will be removed in a future version. Please use the \"create_node\" "
+                "method instead.",
+                DeprecationWarning
+            )
+            call_node = self.call_method("render_wrapper", args, kwargs, lineno=lineno)
+            return self.output(parser, call_node, lineno=lineno, **options)
+
+        return self.create_node(
+            parser,
+            args,
+            kwargs,
+            lineno=lineno,
+            **options
+        )
+
+    def init_parser(self, parser: Parser):
         parser.stream.skip(1)  # skip tag name
 
-    def parse_args(self, parser):
+    def parse_args(
+        self,
+        parser: Parser
+    ) -> Tuple[List[nodes.Expr], List[nodes.Keyword], Dict[str, Any]]:
         args = []
         kwargs = []
+        options = {
+            "target": None
+        }
         require_comma = False
-        target = None
+        arguments_finished = False
 
-        while parser.stream.current.type != 'block_end':
-            if parser.stream.current.test('name:as'):
+        while parser.stream.current.type != "block_end":
+            if parser.stream.current.test("name:as"):
                 parser.stream.skip(1)
-                target = parser.stream.expect('name').value
-                if parser.stream.current.type != 'block_end':
+                options["target"] = parser.stream.expect("name").value
+                arguments_finished = True
+
+            if arguments_finished:
+                if not parser.stream.current.test("block_end"):
                     parser.fail(
-                        'Invalid assignment target', parser.stream.current.lineno
+                        "expected token 'block_end', got {!r}".format(
+                            describe_token(parser.stream.current)
+                        ),
+                        parser.stream.current.lineno
                     )
                 break
 
             if require_comma:
-                parser.stream.expect('comma')
+                parser.stream.expect("comma")
 
                 # support for trailing comma
-                if parser.stream.current.type == 'block_end':
+                if parser.stream.current.type == "block_end":
                     break
 
             if (
-                parser.stream.current.type == 'name'
-                and parser.stream.look().type == 'assign'
+                parser.stream.current.type == "name"
+                and parser.stream.look().type == "assign"
             ):
                 key = parser.stream.current.value
                 parser.stream.skip(2)
@@ -63,41 +100,82 @@ class BaseTemplateTag(Extension):
                 kwargs.append(nodes.Keyword(key, value, lineno=value.lineno))
             else:
                 if kwargs:
-                    parser.fail('Invalid argument syntax', parser.stream.current.lineno)
+                    parser.fail("Invalid argument syntax", parser.stream.current.lineno)
                 args.append(parser.parse_expression())
 
             require_comma = True
 
-        return args, kwargs, target
+        return args, kwargs, options
 
-    def output(self, parser, call_node, target, tag_name, lineno):
+    def create_node(
+        self,
+        parser: Parser,
+        args: List[nodes.Expr],
+        kwargs: List[nodes.Keyword],
+        *,
+        lineno: int,
+        **options
+    ) -> nodes.Node:
         raise NotImplementedError
 
+
+class StandaloneTag(BaseTemplateTag):
+    safe_output: ClassVar[bool] = False
+
+    def create_node(
+        self,
+        parser: Parser,
+        args: List[nodes.Expr],
+        kwargs: List[nodes.Keyword],
+        *,
+        lineno: int,
+        **options
+    ) -> nodes.Node:
+        call_node = self.call_method("render_wrapper", args, kwargs, lineno=lineno)
+        if self.safe_output:
+            call_node = nodes.MarkSafeIfAutoescape(call_node, lineno=lineno)
+
+        if options["target"]:
+            target_node = nodes.Name(options["target"], "store", lineno=lineno)
+            return nodes.Assign(target_node, call_node, lineno=lineno)
+
+        return nodes.Output([call_node], lineno=lineno)
+
     def render_wrapper(self, *args, **kwargs):
-        self.context = kwargs.pop('_context', None)
-        self.template = kwargs.pop('_template', None)
-        self.lineno = kwargs.pop('_lineno', None)
-        self.tag_name = kwargs.pop('_tag_name', None)
+        self.context: Context = kwargs.pop("_context")
+        self.template: str = kwargs.pop("_template")
+        self.lineno: int = kwargs.pop("_lineno")
+        self.tag_name: str = kwargs.pop("_tag_name")
         return self.render(*args, **kwargs)
 
     def render(self, *args, **kwargs):
         raise NotImplementedError
 
 
-class StandaloneTag(BaseTemplateTag):
-    def output(self, parser, call_node, target, tag_name, lineno):
-        if target:
-            target_node = nodes.Name(target, 'store', lineno=lineno)
-            return nodes.Assign(target_node, call_node, lineno=lineno)
-        call = nodes.MarkSafe(call_node, lineno=lineno)
-        return nodes.Output([call], lineno=lineno)
-
-
 class ContainerTag(BaseTemplateTag):
-    def output(self, parser, call_node, target, tag_name, lineno):
-        body = parser.parse_statements(['name:end%s' % tag_name], drop_needle=True)
+    def create_node(
+        self,
+        parser: Parser,
+        args: List[nodes.Expr],
+        kwargs: List[nodes.Keyword],
+        *,
+        lineno: int,
+        **options
+    ) -> nodes.Node:
+        call_node = self.call_method("render_wrapper", args, kwargs, lineno=lineno)
+        body = parser.parse_statements(("name:end%s" % options["tag_name"],), drop_needle=True)
         call_block = nodes.CallBlock(call_node, [], [], body).set_lineno(lineno)
-        if target:
-            target_node = nodes.Name(target, 'store', lineno=lineno)
+        if options["target"]:
+            target_node = nodes.Name(options["target"], "store", lineno=lineno)
             return nodes.AssignBlock(target_node, None, [call_block], lineno=lineno)
         return call_block
+
+    def render_wrapper(self, *args, **kwargs):
+        self.context: Context = kwargs.pop("_context")
+        self.template: str = kwargs.pop("_template")
+        self.lineno: int = kwargs.pop("_lineno")
+        self.tag_name: str = kwargs.pop("_tag_name")
+        return self.render(*args, **kwargs)
+
+    def render(self, *args, **kwargs):
+        raise NotImplementedError